+++ anchor/common/ssv_types/src/share.rs
@@ -5,4 +5,6 @@ use types::PublicKey;
 pub struct Share {
     /// The public key of this Share
     pub share_pubkey: PublicKey,
+    /// The encrypted private key of the share
+    pub encrypted_private_key: [u8; 256],
 }
